% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COQUE-package.R
\docType{package}
\name{COQUE-package}
\alias{COQUE}
\alias{COQUE-package}
\title{COQUE: Simultaneous coefficient clustering and sparsity for multivariate mixed models}
\description{
COQUE is an computationally scalable (at least at the time of development!) approach to performing joint clustering and variable selection of fixed effect coefficients in multivariate longitudinal/independent cluster mixed models. The method works by first fitting separate/stacked univariate longitudinal mixed models to each of the responses (in parallel). The model estimates into a quadratic form resembling that of a composite likelihood multivariate Wald statistic. We then augment the composite quadratic form with broken adaptive ridge regression penalties to encourage sparsity and clustering of the fixed effect coefficients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhui28/COQUE}
  \item Report bugs at \url{https://github.com/fhui28/COQUE/issues}
}

}
\author{
\strong{Maintainer}: Francis K.C. Hui \email{fhui28@gmail.com}

}
\keyword{internal}
