% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\name{estimates}
\alias{estimates}
\alias{estimate.coque}
\alias{estimates.coque}
\title{Extract values from a penalized composite quadratic estimator (COQUE) fit.}
\usage{
estimates(object, ...)

\method{estimates}{coque}(object, lambda, hybrid = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{coque}.}

\item{...}{Not used.}

\item{lambda}{The tuning parameter at which estimates are estimated. This can either be a string taking values e.g., \code{"BIC"} corresponding to taking the estimates based on minimizing of these information criterion, or a numeric value corresponding to specific tuning parameter value. Similar to \code{\link[glmnet:predict.glmnet]{glmnet::coef.glmnet()}}, in the case of the latter it uses linear interpolation to construct parameter estimates for values of \code{lambda} that do not coincide with those used in the fitting algorithm.}

\item{hybrid}{Should a hybrid estimator be calculated also? If \code{TRUE}, then aside from the penalized estimator, a second hybrid estimator is obtained by fitting the composite quadratic approximation subject to the zero constraints implied by the penalized estimator i.e., parameters shrunk to zero are constrained to be zero.}
}
\value{
A list with the following elements:
\item{fixed_effects_matrix}{The matrix of estimated fixed effect coefficients. Due to the use of the broken adaptive ridge (fusion) penalty, some of the estimated coefficients will be shrunk to zero, but coefficients within a covariate many also take the same value i.e., are fused together.}
\item{random_effects_covariance}{A list of random effects covariance matrices for each responses.}
\item{psihat}{The full vector of parameter estimates at the specified tuning parameter value.}
\item{lambda}{The numerical value of the tuning parameter corresponding to the extracted estimates.}
\item{hybrid_fixed_effects_matrix}{If \code{hybrid = TRUE}, then the matrix of hybrid estimator of the fixed effect coefficients. Due to the use of the broken adaptive ridge (fusion) penalty, some of the estimated coefficients will be shrunk to zero, but coefficients within a covariate many also take the same value i.e., are fused together.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts estimates of the fixed effects coefficients and random effects covariance matrix, along wth various other tidbits of information, from a \code{COQUE} fit.
}
\details{
See associated manuscript (currently in revision).
}
\examples{
\dontrun{
# See help function for the main coque() function.
}


}
\seealso{
\code{\link[=coque]{coque()}} for the main function for constructing a COQUE fit, \code{\link[=stackedGLMMs]{stackedGLMMs()}} for fitting stacked univariate generalized linear mixed models, and \code{\link[=generatedata_mglmm]{generatedata_mglmm()}} for simulating data from a longitudinal/independent cluster multivariate generalized linear mixed model.
}
\author{
Francis K.C. Hui \href{mailto:francis.hui@anu.edu.au}{francis.hui@anu.edu.au}
}
