% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{generatedata_mglmm}
\alias{generatedata_mglmm}
\title{Simulate data from a independent cluster multivariate generalized linear mixed model (GLMM).}
\usage{
generatedata_mglmm(
  formula_X,
  formula_Z,
  data,
  fixed_effects,
  Sigma,
  phi = NULL,
  power = NULL,
  trial_size = 1,
  cutoffs = NULL,
  response_type = "gaussian",
  max_count = Inf,
  attempts = 50
)
}
\arguments{
\item{formula_X}{One-sided formula for the fixed effects component of the model. All responses are assumed to be effected by the same formula; sparsity and clustering is instead controlled by the \code{fixed_effects} argument.}

\item{formula_Z}{One-sided formula for the random effects component of the model. All responses are assumed to be effected by the same formula; sparsity and clustering is instead controlled by the \code{Sigma} argument.}

\item{data}{A data frame from which containing relevant fixed and random effect covariate information, and, importantly, \strong{must} also contain a column called "id" which indexes the cluster of each row in the data fame.}

\item{fixed_effects}{A matrix of fixed effect coefficients, where the number of rows is equal to the number of responses and the number columns must be equal to the number of columns of the model matrix implied by \code{formula_X}.}

\item{Sigma}{A random effects covariance matrix, where the number of rows and columns both must equal to the number of responses multiplied by the number of columns of the model matrix implied by \code{formula_Z}.}

\item{phi}{A vector of positive response-specific dispersion parameters, which is needed for Gaussian and Tweedie distributed responses.}

\item{power}{A vector of response-specific power parameters, which is needed for Tweedie distributed responses. All elements should be between 1 and 2.}

\item{trial_size}{A vector of response-specific trial sizes, which is needed for binomial distributed responses.}

\item{cutoffs}{A matrix of response-specific cutoffs, where the number of rows is equal to the number of responses, the number of columns determines the number of levels for the ordinal response, and the first column of the matrix must be a matrix of zeros for parameter identifiability reasons (since a fixed intercept is assumed to be included by default as part of \code{formula_X}).}

\item{response_type}{The distribution of the responses for the model. Currently, it is assumed all responses follow the same distribution, which can be one of "gaussian", "poisson", "binomial", "ordinal", "tweedie".}

\item{max_count}{This argument can be used to try and control the size of the simulated responses. Basically, for any given response the function will try \code{attempts} times to (re)simulate until all the responses are below \code{max_count}. Currently this only appplies to Poisson and Tweedie distributed response.}

\item{attempts}{The number of attempts used to try and re-simulate responses subject to \code{max_count}.}
}
\value{
A list with the following elements:
\item{y:}{The simulated data frame of multivariate longitudinal responses, where the number of columns equal to the number of responses plus one. The last column is just equal to \code{data$id}.}
\item{random_effects:}{A list of simulated random effects with length equal to the number of clusters. Each element in the list is a matrix with the number of rows equal to the number of responses and the number of columns equal to number of columns of the model matrix implied by \code{formula_Z}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Simulates responses from a multivariate longitudinal/independent cluster GLMM, based on a different formulas for the fixed and random effects components, along with response-specific nuisance parameters depending on the response type.
}
\section{Warning}{
Note \strong{no} checks are made on the argument \code{Sigma} to see if it is a positive definite matrix. Please be careful about this!
}

\examples{
\dontrun{
# See help function for the main coque() function.
}


}
\seealso{
\code{\link[=coque]{coque()}} for the main function for constructing a COQUE fit, \code{\link[=estimates.coque]{estimates.coque()}} for extracting values from a COQUE fit, and \code{\link[=stackedGLMMs]{stackedGLMMs()}} for fitting stacked univariate generalized linear mixed models.
}
\author{
Francis K.C. Hui \href{mailto:francis.hui@anu.edu.au}{francis.hui@anu.edu.au}
}
